+++ classes/plugin/Shortcode.php
@@ -2,84 +2,23 @@
 
 namespace Grav\Plugin\ShortcodeCore;
 
-use Grav\Common\Config\Config;
-use Grav\Common\Grav;
-use Grav\Common\Twig\Twig;
-use Thunder\Shortcode\Shortcode\ShortcodeInterface;
+// Check if the new class has been autoloaded. If not, trigger deprecation error.
+if (!class_exists(\Grav\Plugin\Shortcodes\Shortcode::class, false)) {
+    @trigger_error(
+        Shortcode::class . ' class is deprecated, use \\Grav\\Plugin\\Shortcodes\\Shortcode instead',
+        E_USER_DEPRECATED
+    );
+}
 
-class Shortcode
-{
-    /** @var ShortcodeManager */
-    protected $shortcode;
+// Create alias for the deprecated class.
+class_alias(\Grav\Plugin\Shortcodes\Shortcode::class, Shortcode::class);
 
-    /** @var Grav  */
-    protected $grav;
-
-    /** @var Config */
-    protected $config;
-
-    /** @var Twig */
-    protected $twig;
-
+// Make sure that both IDE and composer knows about the deprecated class.
+if (false) {
     /**
-     * Shortcode constructor.
+     * @deprecated 4.2.1 This was a bad idea, reverting back to the old class.
      */
-    public function __construct()
+    abstract class Shortcode extends \Grav\Plugin\Shortcodes\Shortcode
     {
-        $this->grav = Grav::instance();
-        $this->shortcode = $this->grav['shortcode'];
-        $this->config = $this->grav['config'];
-        $this->twig = $this->grav['twig'];
     }
-
-    /**
-     * Initialize shortcode handler
-     */
-    public function init()
-    {
-        $this->shortcode->getHandlers()->add('u', static function(ShortcodeInterface $shortcode) {
-            return $shortcode->getContent();
-        });
-    }
-
-    /**
-     * Returns the name of the class if required
-     * 
-     * @return string the name of the class
-     */
-    public function getName()
-    {
-        return get_class($this);
-    }
-
-    /**
-     * @return string
-     */
-    public function getParser()
-    {
-        return $this->config->get('plugins.shortcode-core.parser');
-    }
-
-    /**
-     * @param ShortcodeInterface $sc
-     * @param string|null $default
-     * @return string|null
-     */
-    public function getBbCode(ShortcodeInterface $sc, $default = null)
-    {
-        $code = $default;
-
-        if ($this->getParser() === 'wordpress') {
-            $params = $sc->getParameters();
-            if (is_array($params)) {
-                $keys = array_keys($params);
-                $code = trim(array_shift($keys), '=');
-            }
-        } else {
-            $code = $sc->getBbCode();
-        }
-
-        return $code;
-    }
-
 }