+++ classes/shortcodes/LanguageShortcode.php
@@ -2,7 +2,6 @@
 namespace Grav\Plugin\Shortcodes;
 
 use Grav\Common\Language\Language;
-use Grav\Plugin\ShortcodeCore\Shortcode;
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
 
 class LanguageShortcode extends Shortcode